+++ CHANGELOG.md
@@ -6,6 +6,29 @@
 
 ### Minor Changes
 
+## 3.5.9
+
+### Minor Changes
+
+- Added Anime Mainnet and Sepolia, Story Mainnet and Aeneid, Megaeth Testnet
+
+## 3.5.8
+
+### Minor Changes
+
+- Added Berachain Bepolia, Celo Baklava, Ethereum Hoodi, Gensyn Testnet, 
+Lens Mainnet, Settlus Mainnet, Superseed Mainnet and Sepolia, Tea Sepolia
+- Deprecated Berachain Bartio
+- Replace `optimism` with `opt-mainnet`
+
+## 3.5.7
+
+### Minor Changes
+
+- Added Solana address activity webhook
+- Added animation to NFT type
+- Added name to create-webhook
+
 ## 3.5.6
 
 ### Minor Changes